refactor: extract the transform logic to crd_transform.go
